Pig Game
Specs asked to code the dice game, pig with two players. If current player rolled a 1, their score would be reset, but other numbers would be added to their total score. Players can choose when to stop rolling or to continue rolling, and whoever reaches a given score first, wins.